tests/requirements.txt
requests==2.31.0
pytest==8.0.0
pytest-asyncio==0.23.5
websockets==12.0
python-dotenv==1.0.1
faker==22.6.0
